project_backend/backend/requirements.txt
aerich==0.8.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncclick==8.1.7.2
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==43.0.0
cycler==0.12.1
dictdiffer==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
fonttools==4.55.0
gmpy2==2.1.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.3
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_resources==6.4.5
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
numpy==1.24.1
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.0.3
parso==0.8.4
passlib==1.7.4
pillow==10.2.0
pip==24.3.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptpython==3.0.29
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
Pygments==2.18.0
pypika-tortoise==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
redis~=3.5.3
requests==2.28.1
rich==13.9.4
rsa==4.9
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tomlkit==0.13.2
torch==1.13.1
torchaudio==0.13.1
tortoise-orm==0.21.7
torchvision==0.14.1
tqdm==4.67.0
typer==0.12.5
typer-slim==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
uvicorn==0.32.1
ultralytics
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
wheel==0.45.1
zipp==3.20.2
